moving genesis_generator tool to genesis folder
- set genesis::startTime to each depositOffer::Start
- omits empty values in platform allocation as in https://github.com/chain4travel/caminogo/pull/124
- includes regenerated genesis_kopernikus file
- no need to adjust the start & end times in deposit offers any more
