+++ scripts/build_genesis_generator.sh
@@ -10,4 +10,4 @@ CAMINO_NODE_PATH=$( cd "$( dirname "${BASH_SOURCE[0]}" )"; cd .. && pwd )
 source "$CAMINO_NODE_PATH"/scripts/constants.sh
 
 target="$build_dir/genesis-generator"
-go build -o "$target" "$CAMINO_NODE_PATH/tools/genesis_generator/"*.go
+go build -o "$target" "$CAMINO_NODE_PATH/tools/genesis/"*.go
