+++ tools/genesis/README.md
@@ -20,6 +20,12 @@ Tool produces a file `genesis_<NETWORK_NAME>.json` in the `OUTPUT_PATH` director
 
 Tool assumes multisignature definitions and allocations are provided in the Excel file. Any other information that shall be contained in the resulting genesis file must be provided in the JSON template.
 
+**ARTIFACTS:**
+
+Inside this folder the following genesis files are stored in subfolders. There are:
+- `templates` - with template file for the generator
+- `generated` - generated files for the networks.
+
 **MORE INFO:**
 
 Just read the code. It's not that long.
