+++ tools/genesis/templates/template_depositOffers_kopernikus.json
@@ -13,87 +13,87 @@
     "depositOffers": [
       {
         "interestRateNominator": 80000,
-        "start": 0,
-        "end": 110376000,
+        "startOffset": 0,
+        "endOffset": 110449241,
         "minAmount": 1,
-        "minDuration": 110376000,
-        "maxDuration": 110376000,
-        "unlockPeriodDuration": 31536000,
-        "noRewardsPeriodDuration": 15768000,
+        "minDuration": 110449241,
+        "maxDuration": 110449241,
+        "unlockPeriodDuration": 31556926,
+        "noRewardsPeriodDuration": 15778463,
         "flags": {
           "locked": true
         }
       },
       {
         "interestRateNominator": 0,
-        "start": 0,
-        "end": 157680000,
+        "startOffset": 0,
+        "endOffset": 157784630,
         "minAmount": 1,
-        "minDuration": 157680000,
-        "maxDuration": 157680000,
-        "unlockPeriodDuration": 63072000,
-        "noRewardsPeriodDuration": 31536000,
+        "minDuration": 157784630,
+        "maxDuration": 157784630,
+        "unlockPeriodDuration": 63113852,
+        "noRewardsPeriodDuration": 31556926,
         "flags": {
           "locked": true
         }
       },
       {
         "interestRateNominator": 90000,
-        "start": 0,
-        "end": 141912000,
+        "startOffset": 0,
+        "endOffset": 142006167,
         "minAmount": 1,
-        "minDuration": 141912000,
-        "maxDuration": 141912000,
-        "unlockPeriodDuration": 31536000,
-        "noRewardsPeriodDuration": 15768000,
+        "minDuration": 142006167,
+        "maxDuration": 142006167,
+        "unlockPeriodDuration": 31556926,
+        "noRewardsPeriodDuration": 15778463,
         "flags": {
           "locked": true
         }
       },
       {
         "interestRateNominator": 100000,
-        "start": 0,
-        "end": 173448000,
+        "startOffset": 0,
+        "endOffset": 173563093,
         "minAmount": 1,
-        "minDuration": 173448000,
-        "maxDuration": 173448000,
-        "unlockPeriodDuration": 31536000,
-        "noRewardsPeriodDuration": 15768000,
+        "minDuration": 173563093,
+        "maxDuration": 173563093,
+        "unlockPeriodDuration": 31556926,
+        "noRewardsPeriodDuration": 15778463,
         "flags": {
           "locked": true
         }
       },
       {
         "interestRateNominator": 90000,
-        "start": 0,
-        "end": 31536000,
+        "startOffset": 0,
+        "endOffset": 31556926,
         "minAmount": 1,
-        "minDuration": 31536000,
-        "maxDuration": 31536000,
-        "unlockPeriodDuration": 31536000,
-        "noRewardsPeriodDuration": 15768000,
+        "minDuration": 31556926,
+        "maxDuration": 31556926,
+        "unlockPeriodDuration": 31556926,
+        "noRewardsPeriodDuration": 15778463,
         "flags": {
           "locked": true
         }
       },
       {
         "interestRateNominator": 110000,
-        "start": 0,
-        "end": 63072000,
+        "startOffset": 0,
+        "endOffset": 63113852,
         "minAmount": 1,
         "minDuration": 60,
-        "maxDuration": 31536000,
+        "maxDuration": 31556926,
         "unlockPeriodDuration": 0,
         "noRewardsPeriodDuration": 0,
         "flags": {}
       },
       {
         "interestRateNominator": 210000,
-        "start": 0,
-        "end": 63072000,
+        "startOffset": 0,
+        "endOffset": 63113852,
         "minAmount": 1,
         "minDuration": 86400,
-        "maxDuration": 31536000,
+        "maxDuration": 31556926,
         "unlockPeriodDuration": 43200,
         "noRewardsPeriodDuration": 43200,
         "flags": {}
